requirements.txt
pdf2image
pytesseract
Pillow
poppler-utils  # sistemine manuel kurman gerekir (PDF dönüştürmek için)
